% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineagespot.R
\name{isGff3}
\alias{isGff3}
\title{isGff3}
\usage{
isGff3(file)
}
\arguments{
\item{file}{Path to GFF3 file.}
}
\value{
result; TRUE if the input file is in GFF3 format, FALSE if not.
}
\description{
Identify whether a file is in GFF3 format.
}
\examples{

gff3_path <- system.file("extdata","NC_045512.2_annot.gff3",
                         package = "lineagespot")
isGff3(gff3_path)

}
