% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineagespot.R
\name{isVcf}
\alias{isVcf}
\title{isVcf}
\usage{
isVcf(file)
}
\arguments{
\item{file}{A path to a folder containing all VCF files
that will be integrated into a single table (or A character vector of paths
to VCF files.)}
}
\value{
A list of two elements;
\itemize{
\item VCF list; A list where only VCF files are stored.
\item number of VCF files; A parameter indicating the number of VCF files
stored in the VCF list.
}
}
\description{
Identify VCF files from a group of files.
}
\examples{

vcf_exists <- isVcf(system.file("extdata", "vcf-files",
                                package = "lineagespot"))
print(vcf_exists)

}
