% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checking.R
\name{list_gff3}
\alias{list_gff3}
\title{list_gff3}
\usage{
list_gff3(file)
}
\arguments{
\item{file}{Path to GFF3 file.}
}
\value{
result; TRUE if the input file is in GFF3 format, FALSE if not.
}
\description{
Identify whether a file is in GFF3 format.
}
\examples{

gff3_path <- system.file("extdata", "NC_045512.2_annot.gff3",
  package = "lineagespot"
)
list_gff3(gff3_path)
}
