% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_info_hits.R
\name{outbreak_info_hits}
\alias{outbreak_info_hits}
\title{outbreak_info_hits}
\usage{
outbreak_info_hits(vcf_table = NULL, ref_folder = NULL, print.out = FALSE)
}
\arguments{
\item{vcf_table}{A tab-delimited table containing all variants for all samples}

\item{ref_folder}{A path to outbreak.info reports}

\item{print.out}{logical value indicating if the produced table should be printed}
}
\value{

}
\description{
Find overlapping variants with SARS-CoV-2 reference lineages
coming from outbreak.info reports
}
