% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineage_report.R
\name{get_lineage_report}
\alias{get_lineage_report}
\title{Title}
\usage{
get_lineage_report(
  lineages,
  base.url = "https://api.outbreak.info/genomics/lineage-mutations?pangolin_lineage="
)
}
\arguments{
\item{lineages}{a character vector containing the names of the lineages of interest}

\item{base.url}{The base API URL used to search for lineage reports}
}
\value{
List
}
\description{
Title
}
\examples{


}
