% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checking.R
\name{isVcf}
\alias{isVcf}
\title{isVcf}
\usage{
isVcf(vcf_fls = NULL, vcf_folder = NULL, gff3_path = NULL, vcf_exists = 0)
}
\arguments{
\item{vcf_fls}{A character vector of paths to VCF files}

\item{vcf_folder}{A path to a folder containing all VCF files
that will be integrated into a single table}

\item{gff3_path}{Path to GFF3 file containing SARS-CoV-2 gene coordinates.}

\item{vcf_exists}{A parameter indicating the number of VCF files
stored in the VCF list.}
}
\value{
A list of two elements;
\itemize{
\item VCF list; A list where only VCF files are stored.
\item number of VCF files; A parameter indicating the number of VCF files
stored in the VCF list.
}
}
\description{
Identify VCF files from a group of files.
}
\examples{

isVcf_info <- isVcf(
    vcf_folder = system.file("extdata", "vcf-files",
                             package = "lineagespot"
    ),
    gff3_path = system.file("extdata",
                          "NC_045512.2_annot.gff3",
                          package = "lineagespot"
    )
)
print(isVcf_info)

}
