% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineagespot.R
\name{lineagespot}
\alias{lineagespot}
\title{lineagespot function}
\usage{
lineagespot(
  vcf_fls = NULL,
  vcf_folder = NULL,
  gff3_path = NULL,
  ref_folder = NULL,
  voc = c("B.1.617.2", "B.1.1.7", "B.1.351", "P.1"),
  AF_threshold = 0.8,
  file.out.index = Sys.Date(),
  print.out = FALSE
)
}
\arguments{
\item{vcf_fls}{A list of paths to vcf files}

\item{vcf_folder}{A path to a folder containing all VCF file that will be integrated into a single table}

\item{gff3_path}{Path to gff file}

\item{ref_folder}{A path to outbreak.info reports}

\item{voc}{a character vector containing the names of the lineages of interest}

\item{AF_threshold}{A parameter indicating the AF threshold for identifying variants per sample}

\item{file.out.index}{A string index merged at the output files}

\item{print.out}{logical value indicating if the produced table should be printed}
}
\value{
List
}
\description{
Identify SARS-CoV-2 related mutations based on a single (or a list) of
variant(s) file(s)
}
