% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineagespot_hits.R
\name{lineagespot_hits}
\alias{lineagespot_hits}
\title{outbreak_info_hits}
\usage{
lineagespot_hits(
  vcf_table = NULL,
  ref_folder = NULL,
  file.out = paste0("lineage_hits_", Sys.Date(), ".txt"),
  print.out = FALSE
)
}
\arguments{
\item{vcf_table}{A tab-delimited table containing all variants for all samples}

\item{ref_folder}{A path to outbreak.info reports}

\item{file.out}{Given name for the output file}

\item{print.out}{logical value indicating if the produced table should be printed}
}
\value{

}
\description{
Find overlapping variants with SARS-CoV-2 reference lineages
coming from outbreak.info reports
}
