% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vcf.R
\name{merge_vcf}
\alias{merge_vcf}
\title{merge_vcf}
\usage{
merge_vcf(vcf_fls = NULL, vcf_folder = NULL, print.out = FALSE)
}
\arguments{
\item{vcf_fls}{A list of paths to vcf files}

\item{vcf_folder}{A path to a folder containing all VCF file that will be integrated into a single table}

\item{print.out}{logical value indicating if the produced table should be printed}
}
\value{

}
\description{
Merge Variant Calling Format (VCF) files into a single tab-delimited table
}
