% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq_variants.R
\name{uniq_variants}
\alias{uniq_variants}
\title{Lineage report for variants overlapping}
\usage{
uniq_variants(
  hits_table = NULL,
  AF_threshold = 0.8,
  file.out = paste0("lineage_report_", Sys.Date(), ".txt"),
  print.out = FALSE
)
}
\arguments{
\item{hits_table}{A tab-delimited table containing the overlaps identified}

\item{AF_threshold}{A parameter indicating the AF threshold for identifying variants per sample}

\item{file.out}{Given name for the output file}

\item{print.out}{logical value indicating if the produced table should be printed}
}
\value{

}
\description{
Lineage report for variants overlapping
}
